SSG-12617: handle having no images to remove
git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@45068 219f1c6e-4a0e-43c8-96f2-6e749be13fd4

+++ etc/docker/build-image.sh
@@ -13,6 +13,10 @@ removeExistingImages() {
 	echo "Removing existing Docker images from the build agent"
 	echo "Existing images:"
 	docker images
+	if [ $(docker images -q | wc -l) -eq 0 ]; then
+		echo "Done removing existing Docker images from the build agent (no images found)"
+		return
+	fi
 	docker images -q | uniq | xargs docker rmi -f 
 	if [ $(docker images -q | wc -l) -gt 0 ]; then
 		echo "Failed to remove the existing Docker images from the build agent"
